feat(diff): introduce swap dominance percentage for palette swaps
- Added `swap_dominance_pct` to `CostModel` and `SpecOverrides` to control the minimum percentage of source token changes required for a palette swap.
- Updated `collapse_swaps` to utilize the new dominance threshold when determining swaps.
- Enhanced `Diff` struct to include a `support` field representing alignment confidence.
- Refactored `diff_at_raw` to avoid redundant comparisons and improve performance.
- Updated tests to validate the new swap dominance behavior and ensure deterministic output for palette swaps.
- Adjusted JSON serialization to reflect changes in the `Diff` structure.
- Modified FFI and Python bindings to expose the new `swap_dominance_pct` parameter.
- Ensured air blocks are treated as absence in diff calculations to maintain consistency.

+++ nucleation-jvm/jvm/src/main/java/com/github/schemat/nucleation/Diff.java
@@ -59,7 +59,7 @@ public final class Diff implements AutoCloseable {
     }
 
     /** Total edit distance (number of weighted operations) between the two schematics. */
-    public int distance() {
+    public long distance() {
         checkOpen();
         return NucleationNative.nDiffDistance(handle);
     }
